Proto,tests: Add an optional argument [check_size] to [Block.bake]
The development of Transaction Rollups has proven that sometimes, it
is mandatory to check that the arguments of a layer-1 operations
actually fit in a Tezos operation.

This patch improves the [Block.bake] helpers function to do just
that. The [check_size] optional argument is set to [true] by default,
but can be opt-out explicitly.