Lib_lwt_result_stdlib: reduce test count for performance reasons
As described in https://gitlab.com/tezos/tezos/-/merge_requests/5230#note_947965542,
the migration from QCheck to QCheck2 causes a performance hit in the
runtime of the test suite, the most important being `test_fuzzing_list`,
which goes from 3s to 14s after the migration.

This commit tries to alleviate that performance loss, by doing half as
many iterations of each test (50 instead of 100).