Proto: check [sc_rollup_max_lookahead_in_blocks] constant
This constant gives the maximum number of lookahead in blocks during
a commitment submission. The commitment must be posted with a
frequency defined in [sc_rollup_frequency_in_blocks], if the max
lookahead in blocks is lower than the frequency, the commitments will
always be rejected as the lookahead will be at least the frequency.