Docker: Include the rollup binaries in the Octez docker images
As of June 2022, Octez does not provide production-ready binaries for
the two existing solutions of rollup that are currently baked in
Tezos, that is transaction and smart-contract rollups.  The binaries
that are currently under (various degree of) development are not ready
for production use.  As a consequence, they are not included in the
Docker images that we build from our CI infrastructure.

This decision was made to avoid any confusion, to be explicit on the
scope, state, and readiness of our rollup binaries, to avoid anyone
starting using them with the wrong impression.

However, this does pose an additional challenge for rollup testers,
who need to build the binaries they need from source, or for Tezos
core developers, who need to provide dedicated Docker images.

This patch proposes a middle-ground approach, where the binaries are
in fact included in the images built from the CI, but none of the
tooling built on top of these images use them, especially the
`tezos-docker-manager.sh' script.  Rollup testers will have to use the
`bare' image generated by the CI, to start a container and log into
it.