Proto: do not decode scripts in precheck
Precheck no longer tries to deserialize scripts. It does still
check that the operation has enough gas for these deserializations
(by consuming an estimated gas cost based on the bytes size: this
has not changed).

This is a semantic change: if the deserialization fails, the
operation will now pass precheck but its application will fail.
This means that the operation may be included in a block, in
which case the fees will be taken but the operation will have
no further effects.