Proto/tests: fix operation integration tests
With the 1M restriction on the protocol side this tests failed. This
commit replace some [Incremental.add_operation] with [Block.bake].

Most of this calls used incremental construction instead of block
application to create a context with managers and specific
balance/contracts/etc. Block application should be used to
create a specific context and avoid raising 1M restriction error.