CI: fix usage of find in snapshot_alpha.sh
On some CI runners, the script `snapshot_alpha.sh` timeouts. See for
instance:

  https://gitlab.com/g.b.fefe/tezos/-/jobs/2612767802

After some instrumentation (a.k.a. `set -x`), it appears to be stuck
on a call to:

```
find . -type f -exec sed -i.old ...
```

And after some investigation on the runner it appears to be in a
infinite loop where the file created by `sed` are provided back to
find...

While I do not understand why, this patch fix the CI on my own
runners...