refactor: reverse dependency between lib_context and lib_shell_services
This commit moves all types relative to the definition and
transportation of merkle proofs for the light mode, from
`lib_shell_services` to `lib_context`.

This refactoring reverses the dependency between both library: now,
`lib_shell_services` depends on `lib_context`, instead of the reverse.
This will allow the former to depend on other types defined in the
latter, which is a necessary step to achieve #2607.