Proto,wasm: Unplug WASM tests in K
We are making breaking changes in the WASM PVM compared to what has
been frozen in K.  This is OKay, because the WASM PVM is not in
production yet.  The simplest course of action is to remove the faulty
tests, rather to needlessly pollute the code to remain backward
compatible with something that was never intended to be used.