Proto,Scoru: Remove useless number_of_messages from commitments
We now use [max_number_of_messages_per_commitment_period] to limit
the number of messages in an inbox, to limit the inclusion proofs.
This was before limited by the number of available messages which
we also remove, since it can not be computed without the number of
messages in commitments.