WASM: rewrite Chunked_byte_vector.of_string without dependency on Lwt
Rewrites `Chunked_byte_vector.of_string` to no longer rely on the effect. This
avoids introducing Lwt in the text parser when builder the Lwt variant, and
there is no need for `of_string` to be done lazily: it is always a monolithic
transformation that happens in the parser only. It is needs to be done lazily,
it needs to be tickified.