Lib_openapi: fix null handling with Combine constructor
This commit fixes a bug where Null elements in Json schemas would not be
processed correctly when being translated to OpenAPI schemas.

The version of OpenAPI we use does not handle null elements in
disjunctions, so instead we use a `nullable` property on said
disjunctions, and don't convert null elements from Json schemas to
OpenAPI.

The problem was that null elements in disjunctions (i.e. the `Combine`
element kind) can be represented either by an element with the `Null`
kind, or by the following object:
```
{ /* None */
  "none": null }
```
This commit adds the correct processing of the latter representation to
the `convert_element` function.