Tezt: revamp the run_batched_operation test in prevalidator.ml
Remove the run_operation part of this test: indeed, run_operation
is a plugin RPC while prevalidator.ml is about the mempool. Moreover,
this part is now tested in the new
Run_operation_RPC.test_correct_batch test.

Also rename the test to test_inject_manager_batch, clean it up,
and move it to the Revamped module.