Proto: forbid a delegate to propose the same protocol hash twice
This a semantic change: the new Already_proposed error is now
returned when a Proposals operation contains a proposal that
has already been proposed by the delegate.

The goal is to prevent replaying a Proposals operation.

Moreover, for the sake of consistency, if the proposals contain
a duplicate element, the operation is now rejected (with the
new Proposals_contain_duplicate permanent error).