Scoru_wasm: Drop [Thunk], use chunked byte vector in [Gather_floppies]
In order to reduce the overall complexity of the WASM PVM, we drop
[Thunk] in favor of the now usable [Tree_encoding_decoding] framework.

We update [Gather_floppies] accordingly.  The resulting refactoring
gives a more functional flavor to the module.  We also use this
opportunity to reuse the [Chunked_byte_vector] module from
[Lib_webassembly] to store the kernel.

This new implementation is likely not to produce larger proofs. This
is because the steps of the [Gather_floppies] were mostly using the
same fields.