Proto/lib_delegate: VDF daemon injects operation at correct level
Before, the main loop operated on the stream of blocks. The lengthy
VDF computation blocks this loop and all the blocks added to the chain
in that interval are consumed at once after it is finished. This resulted
in the reveal operation being based on the block after the one where the
computation started, which, for high difficulties, would be too old.
The stream is now reinitialised after the VDF computation.