Tezt,SCORU: Test catch up mechanism of rollup node
We introduce a (failing) test for the synchronization mechanism of the
rollup node. The test originates a rollup, bakes 65 blocks, and
launches the rollup node.

When it is catching up with these first 65 blocks, the rollup node
needs to publish two commitments. The problem is that it does that so
fast that the layer 1 does not have the time to inject the first
commitment in a block before the second commitment is considered for
publication by the rollup node.

For this reason, the simulation of this second publication fails,
and the rollup node stops.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>