SCORU,Node: Check consistency with layer 1 inbox
The rollup node maintains its own inbox because this inbox contains
all the history required to produce inclusion proofs for refutation
games.

It is critical that the inbox descriptor of the rollup node stays
consistent with the one computed by the layer 1 or else the refutation
game will be lost by the rollup node.

This invariant is unfortunately invalid: adding this check in the
rollup node makes the integration tests fail, witnessing a bug in
the rollup node.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>