Proto/Scoru: Parse wasm bootsector from Hex value
The 'parse_boot_sector' function, albeit defined in the protocol, in
only used by the 'tezos-client' to parse the boot sector provided as
its command line argument.

Since the WASM boot sector is expected to be of binary format, the
current implementation is too fragile, and we propose to modify it to
make so 'tezos-client' expect a hex encoded string instead.