Tree_encoding: Reorganize the dependencie tree of 'lib_scoru_wasm'
Before to this patch, the dependency tree of lib_scoru_wasm is
straightforward.

    lib_webassembly         <- lib_scoru_wasm
    (provide lazy containers)  (provide tree encoding)

The caveat is that the encoding of the types introduced in
lib_webassembly are defined in lib_scoru_wasm… and therefore cannot be
easily tested in the official WASM test-suite (in a decode, eval,
encode loop).

This patch is the first step towards being able to test the tree
encoders developed for the PVM in the WASM test-suite.

We introduce two libraries,

  - 'lazy-containers' introduces laz maps, lazy vectors, and chunked
    byte vectors
  - 'tree-encoding' introduces the tree encoders (including for the
    lazy containers)

The rest of the code base is unchanged, in particular the tree
encoding are still defined in 'lib_scoru_wasm'.