test: improve Account.generate_accounts use
`generate_accounts` were used to generate the list use for bootstrap
account. I make that function only generating accounts and created a new
function that make bootstrap account type from a list of account, with
?initial_balances and ?boostrap_delegations

I also simplified the use of `boostrap_delegations` with the same
logic os `initial_balances`. Which is the n account is delegated to the
n element of the list `bootstrap_delegations`.