Proto,SCORU: Increase max number of messages per commitment period
Fixes https://gitlab.com/tezos/tezos/-/issues/2373

The value is chosen to limit the length of inbox refutation proofs. In
the worst case, the length of inbox refutation proofs are
logarithmic (in basis 2) in the number of messages in the inboxes
during the commitment period.

With the following value, an inbox refutation proof is made of at most
35 hashes, hence a payload bounded by 35 * 48 bytes, which far below
than the 32kb of a Tezos

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>