doc: add temporary fix and warnings to build from scratch doc recipe
Fixes !3592

The changes brought forward by tezos/tezos!5027 had a weird
interaction for users which had exported `$HOME/tezos` or a similar
folder to their `PATH` environment variable -- as previously suggested
in the docs before tezos/tezos!5278 updated them.

We update the instructions with a command which removes that entry in
the path if present, add a warning, and clearly marks these steps as
optional.

EDIT: Thanks @abate for the improved cleanup recipe.