WASM: Fix tickification of MKElaborateFunc in the parser
The [MKElaborateFunc] tick was unbound, because it was loading every
instructions of a given block to make further checks.

To deal with that, we need to modify [MKElaborateFunc] to check one
instruction at a time, which looks easy enough until you remember some
of these instructions (namely, [Block], [Loop] and [If]) actually
takes instructions as their arguments. This is why we add to
[MKElaborateFunc] a queue of lazy iterators ([LazyVec]), itself
traverse with a lazy iterator ([LazyVec]).

As a consequence, the [MKElaborateFunc] ticks are divided even
further, and the size of each bound should be bounded.