WASM: Clean-up the "wasm" directory
The "wasm" directory in the Irmin tree of the PVM is used to maintain
a lot of information across tick, but as of now, it was never cleaned
up. It meant that the tree would become crowded with data that were no
longer needed.

It turns out it is very simple to clean up, thanks to subtree move.
Because we keep in the lazy-containers the original subtree, and we
inject it prior to updating read keys, the tree-encoding library does
not rely on the input tree at encoding time.

With this, we gain an additional 5% of proof size in the worst tick of
the computation.wasm kernel.