Tezt,WASM: Enable back tezts running kernels as long tests
The WASM PVM is becoming increasingly slow, due to ticks becoming
smaller and smaller to enable small proofs. As a consequence, the
rollup node is becoming slower too, and the tezt tests using the WASM
PVM are taking a longer time to complete.

Because the runners are on average 10 time slower than the typical
developer computer, making these tests part of the CI run for every MR
is not practical (it took more than 30s for certain inbox level to be
computed).

As a consequence, we move these tests to Tezt long tests
infrastructure. These tests are still run, but asynchronously.