proto/scoru: proof_stop_state always returns stop hash
The caller `game_repr` is the one that must check that the input is
valid with regards to the input_request.

The proof can be valid even if the input is no related to the
input_request when the PVM is stuck or if it's waiting for an input that
does not exists yet. In that case the checked stop hash by
`check_proof_start_stop` must be None.