Consensus key documentation
The consensus key is introduced in proto L. Release notes already have
details about the new operations and rpc commands:

https://tezos.gitlab.io/protocols/alpha.html#consensus-key

This MR introduces more genric documentation to help delegates use the
feature. This belongs to the "Key management" section.

This should be merged later, when protocol L is voted in.