Lazy_containers: [produce_value] is for tree-encoding
The [create] function of the [Chunked_byte_vector] data-structure was
providing a default value for [produce_value], and the WASM
interpreter was relying on this default value when creating/allocating
new chunked vector. This works well as long as you do not throw a
decoding/encoding loop in the mix. As soon as you do that, the default
[produce_value] is lost.

This was the result of a context hash split between one N-ticks and N
one-tick executions.