Tree_encoding: Lazy ~default argument for tagged_union
Tree encodings, just like Data encodings, are top-level definitions
that are being initialized at the start-up time of the program that
relies on them.

The [tagged_union] combinator of tree-encoding features a ~default
labeled argument which acts as a fallback when the decoding of a value
fails. This argument is therefore evaluated at initialization time
too.

The WASM PVM uses this labeled argument to encode that, for an empty
tree, the initial state is [Decode]. The initial state of the decode
phase contains mutable fields, that are modified throughout the
parsing of a WASM program.

The consequences of these three facts is that, when running twice a 1
N-tick execution of the first [kernel_next] invocation, the initial
decoding state of the second will be the result of the parsing phase
of the first, meaning two PVM executions from the same tree in the
same process leads to two different context hashes.  Similarly, though
less problematic, each subsequent [kernel_next] parsing phase
interferes with the next.

This patch changes the nature of ~default to take a lazy value,
encoded as a [unit -> 'a] function.