WASM: Fix IK_Eval step
TL;DR: We remove a useless argument from some constructors, and fix a
bug in the process.

This is yet again another bug that was only showing up on 1 N-ticks
execution, but made N 1-tick execution impossible.

The evaluation of a WASM program modifies a so-called module instance
([module_inst]), which notably contains the memories and tables the
module manipulates. During the initialization ([Eval.init]), the final
step is the execution of previously gathered code. This execution is
likely to modify the [module_inst] value of the module, built during
the initialization.

The issue is that, in addition to the [config] that allows a
step-by-step execution, the [IK_Eval] constructor was also taking an
additional argument: the initial value of the [module_inst].

In a 1 N-tick execution, the two values (the one modified by the
evaluation, and the one saved in [IK_Eval]) are actually “the same”
thanks (or due to) the heavy use of mutable variables. As a
consequence, when the evaluation of WASM caused the modification of
the [module_inst], this also modifies the one in [IK_Eval].

This is no longer true in the world of N 1-tick executions: after the
first [encoding] of the [IK_Eval] tick state, the sharing is lost. As
a consequence, a divergence occurs at the first tick that modifies the
contents of the [module_inst].

The best part of this story is: the [module_inst] argument of
[IK_Eval] is actually never used.  So we can fix this bug by simply
removing it, and the one in [IK_Stop] while we are at it.