Lib_context: add a With_get_data functor
This new functor in `src/lib_context/sigs/context.ml` gives a `get_data`
function to any module that has the `Storelike` interface (basically,
any store we use will have that interface). The partial application
`get_data leaf_kind keys` has the correct signature to be used in
Irmin's `{produce,verify}_proof`, and corresponds to a proof that a tree
contains the keys given in argument.