Plugin/mempool: no longer retrieve info from Partial_construction mode
Instead, store the proposal_round in the filter state.

This will allow us to remove the predecessor_round field from the
Partial_construction mode of the application_state in the next commit.
And it prepares for the full removal of the application_state from the
mempool plugin (keeping only the validation_state) that will come with
the environment rework.