Doc: remove phony target: main
This MR reworks how symbolic links are created and removed.
Instead of removing the links just before they are created,
the links are removed before building. This breaks the test
of docs/user/node-config.sh because in parallel of this test,
'make all' is running, and it removes the links.

This actually shows that this test was flaky before, although
it had a very low chance of failing for this particular reason.

This commit fixes that by removing the dependency from doc targets
to the main Octez targets. One now has to build Octez before using
the doc Makefile if one wants the doc to be up-to-date.