Proto+Plugin/Michelson: enrich execution traces with "just consumed gas"
Before this patch, Michelson execution traces contain at each
execution step the amount of remaining gas at the given step. This
commit adds the difference between this amount and the last reported
one, or in other words the amount of gas that was consumed during the
execution of the last step (or everything happening before we start
executing the script in the case of the very first line of the trace).

This is both more informative for the user (who can more easily see
which steps consumes more or less gas) and more stable for regression
testing (only the lines corresponding to changes in gas costs appear
in the diff).

Moreover, the special case of "just consumed gas: 0" hints at gas-free
instructions which we should probably avoid in general.