Shell: duplicate mempool dependencies
This commit just copies the relevant files as they are. The next
commit will modify them into a coherent legacy mempool implementation.
This commit separation is intended to make future rebases easier: when
one of the copied file is modified on master, simply copy it again in
this commit, then handle any conflicts in the subsequent commits.