SCORU/Wasm: (and EVM on WASM) test store_move and store_copy and fix move_tree implementation
Co-Authored-By: Emma Turner <emma.turner@trili.tech>

Previously, store_move was implemented as store_copy followed by delete.
This worked for all cases except where the tree was moved to a key
'inside' itself, in which case the new location of the tree was deleted
along with the old locatation.

Therefore, store_move is amended to perform the deletion of the old location
prior to inserting the tree at the new location.