WASM: Deduplicate the [module_reg] during the init phase
Shared mutability strikes again.

The [module_reg] type is a mutable map of mutable values. During the
[IK_Admin_join] final step, the [module_reg] field of the [Init] tick
state was put in the [config] constructed to evaluate the _start
function.

If it happened that the [module_reg] value was modified during this
evaluation, e.g., in case of a call to the [TableSet] instruction as
it was done during the initialization of the pipe.wasm kernel written
in Rust, then either both [module_reg] are modified (in [Init] and in
the [config]), or only one (in [config]) is modified, depending on
whether or not the shared mutability has been broken by an
encode-decode cycle.

Even worth, the tick-by-tick execution would effectively be broken,
since the [module_reg] instance that is actually modified (in
[config]) is dropped at the end of the execution of _start.

This patch fixes this issue by making sure there is only one
[module_reg] living in the tree.