Tezt/dal: remove restriction to protocol alpha
Most of these "supports:(From_protocol (number Alpha))" date from
June or July, so the "Alpha" they refer to should actually be
Kathmandu or Lima.

With the split of the legacy/new mempools, it becomes even more
important to run this kind of tests both on Kathmandu, to ensure
that the legacy mempool is not broken, and on Lima/Alpha for the
new mempool.