Dal/Scoru: Dal skip lists only persisted for levels >= origin. lvl + lag
The rollup node is not supposed to care about the content of blocks that
are baked before its origination level.

However, the way we currently retrieve and update Dal skip list & cache
from the node's store imposes to track blocks cofirmed at
`levels >= origination level` but published at `levels < origination level`.

Since Dal's refutation game rejects any Dal slot that is published at
a `level <= origination level`, we can safely relax the invariants of
Dal slots history skip list and cache as follows:
- Before `originaton level + endorsement_lag`: just read the values from L1
(for the skip list) or use the default one (for the bounded cache);
- Starting from `originaton level + endorsement_lag`, maintain a correct
vision of the skip list and of the bounded cache in the rollup.