SCORU,WASM: Correctly align ticks number for commitment periods
Previously, the tick to commit was off-by-one, meaning it was not a
multiplier of [MAX_TICKS]. This was only visible on the first
commitment period, since after that, the tick numbers are relative to
the previous period.

This fix will make easier to constrains the tick numbers in a
dissection wrt. snapshot size.