WASM: Fix the error handling of the reveal functions
With this patch, they are now consistent with the other host
functions.  We do that by giving control to the PVM on how the result
of an input tick is loaded into memory. Now, the PVM not only decides
what to do when the [reveal_info] fails, but also when the “push the
result to memory” fails.

It is also a good news for the fast execution, because it allows to
functorize the loading of the result of the tick into the memory, as
it was done for the other host functions.