WASM: Reset the input buffer when the kernel yields
To do that, we had to modify tree_encoding to properly clean-up the
tree when encoding a fresh map. This drastically reduces the cost of
installing a fresh map in a location already used by a map in a
previous iteration (divided by 5). This is logic, because we don't
need the contents of this previous map to construct the hash of the
new subtree.