Scripts: move set -u from build_static_binaries.sh
set -u causes the script to fail if $EXECUTABLE_FILES or $ARCH is unset.
But it's not enough: the script should also fail if $EXECUTABLE_FILES
or $ARCH is empty. So we add an explicit check at the beginning of the
script. This improves the error message in case a variable is unset.