WASM/PVM: fix outbox not created for a level
Basically, the outbox was created during `read_input`. If read_input is never
called, the outbox is never created. Worse than that, write_output always push
messages in the last outbox, which should be the same level. If your read inputs
at a level `n`, then you don't read them at level `n+1` but output messages,
they will be placed in the outbox at level `n`.

We implemented time travel, but I'm not sure the world is ready for that.