Proto: Fix the WASM PVM requests an unknown preimage for invalid hashes
Prior to this patch, the PVM defaults to ask for the preimage of the
zeroed hash, which is unknown as far as we know.

We change its behavior to request the preimage of a well known hash,
so that rollup whose kernel asks for an invalid hash could continue to
progress.