Manifest: bump some minimal version requirements
These dependency version bumps correspond to inter-dependency
dependencies. Specifically, if we depend on a package both directly and
transitively, and if the transitive dependency has a higher lower-bound
version constraint, then we effectively depend on this package with the
higher lower-bound.

Note that in some cases, we depend on a package directly for one of our
internal lib, and transitively for a different one. So we could leave
the lower lower-bound constraint to be less restrictive to users of the
directly-dependent lib. But we don't because that would be more
complicated to work with.