WASM: Merge the level and id arguments of [read_input]
This is to reduces the number of memory pages accessed by the host
function in the worst case.

We had to update the kernel more substantially than just the
`read_input` declaration, because (1) the handling of external
messages was broken, but (2) a very old kernel prior to this
regression was used in Tezt.

Co-authored-by: Pierrick Couderc <pierrick.couderc@nomadic-labs.com>