Scoru/Wasm VM: continue computation when errors have fallback
Previously, evaluation stopped when a Stuck state was reached. If the
error was recoverable through the fallback mechanism, then the
following evaluation would 'unstuck' the vm.  This worked fine for the
PVM, but resulted in the node halting execution when such an error was
reached.

Instead, we continue computation when an error is recoverable.