Proto/Michelson/Gas: slightly bump some costs
Thanks to continuous benchmarking (see #4036), we were able to measure
more precisely many small gas parameters. A few of them showed
irregular timings. We bump them slightly for safety but to remove this
flakiness the corresponding benchmarks should probably be modified to
use amplification (or to increase the number of iterations if
amplification is already used).