Snoop,Scoru: Uniform sample of the log of the skip list len
This commit updates the samplers used for the benchmarks of the skip
list data structure used in Smart Rollups. The models of the
operations on this data structure are logarithmic but the previous
samplers were uniform in the length (in some interval). Because of
this, during the linear inference of the model, very few points of
small abscissa were available and this degraded the precision of the
inferred gas parameters.

To solve this, we now sample the logarithm of the size uniformly
instead of the size itself.