Proto/Doc: avoid refering to contract code as "source code"
The source code of a contract usually means a piece of code written in
a high-level language (such as SmartPy, Ligo, Archetype, or Lorentz)
and compiled to Michelson prior to the origination of the
contract. This piece of code is totally unknown to the protocol.

A few occurrences of the expression "source code" appear in protocol
comments though but they refer to something else: the unparsed
contract script.

This commit edits these comments to avoid confusion.