Event-logging: fix documentation
Error message:
```
Reference to 'level' is ambiguous. Please specify its kind: type-level, val-level.
```

Error:
There are two different kind of things that the `{!level}` reference can
refer to: the `level` type and the 'level` value. In this case it was
the type and so the reference needed to be replaced by `{!type-level}`.