Micheline: fix documentation
Error message:
```
File "src/lib_micheline/test/test_diff.ml", line 29, characters 29-33:
Error: Unknown tag '@src'.
```

Error:
The form `Invocation:   dune build @src/lib_micheline/runtest` is meant
to be read by the devs but not by the machines. The `@` has a different
meaning there than in docstrings.

There are two ways to fix this: escape the string (with `[...]` markup),
make the comment not a docstring (using a single-asterisk comment
`(* ... *)`). I went for the second because there is no reason.

Note that the same form is used in other files. They don't trigger
errors because they are test targets (hence executables) but these tests
here are written as libraries.