Clic: fix documentation error
Error message:
```
File "src/lib_clic/unix/scriptable.mli", line 34, characters 30-51:
Error: Reference to 'command' is ambiguous. Please specify its kind: type-command, val-command.
```

Error:
The reference could be to either of two different kinds of identifiers:
type or value. In this case it's a the type and so the `type-` prefix
was needed.