Baker: move confusing event placement
If the round start (of the first potential round at next level) is in the past,
the event will show something like
```
 waiting -38min15s until it's time to bake at ...
```
which is quite confusing. Actually the baker does not wait.

So the baker now emits this event only when it really waits.

Fixes https://gitlab.com/tezos/tezos/-/issues/4646