Shell/test_prevalidation: provide proto_outcome as part of the state
instead of as validation_info. Indeed, it is not more difficult, and
it is more similar to what we will be doing in the future with the
bounding_state.

We also rename get_valid_operations to get_mempool_operations to
stress that they are the operations recorded by the protocol mempool,
not to be confused with the records by other states (filter_state for
now, bounder_state in the future).