Proto/Token: rename "source" into "giver"
There is a risk of confusion between the notion of transaction
source (the account signing and/or sending a transaction) and the
notion of token transfer source (the account being debited). Most of
the time this is not a problem because the accounts coincide but
sometimes it is good to explicit the distinction.