Proto/Michelson: internal renaming of "union" into "or"
The Michelson type "or" is referred to as "union" in the protocol to
avoid clashing with OCaml's keyword "or".

This is inconsistent with the way all other Michelson types are
treated.

This commit renames all occurrences of "union" into "or" except the
few ones that would clash with the OCaml keyword (the definition of
the type used to interpret "or" in the IR: Script_typed_ir.union and
two functions to construct "or" types outside of the protocol:
Types.union in lib_benchmarks/lib_benchmark_type_inference and
Michelson_types.union in lib_benchmarks_proto) which are renamed to
"or_" instead.

Co-authored-by: Julien Tesson <julien.tesson@nomadic-labs.com>