Proto: decrease the constant gas cost of manager operation
The 1000 value for this constant was mostly there to cover for the
cost of checking signatures which is now handled more directly; with
the previous commits, the cost now depends on the signature scheme
being used and the length of the operation but more importantly it is
only consumed as many times as signatures are indeed checked (only
once for each batch and never for internal operations). The remaining
100 cost is there to cover the resources used in apply either
directly (dispatching on operation kinds) or indirectly (in particular
in the production of operation results).