DAC: Add optional `dac_node_cctxt` to the legacy mode
Optional `dac_node_cctxt` field was added to the configuration
of the dac node running in the `legacy` mode. This way we would
be able to run end to end tests, where one dac node will mimic
coordinator and another will mimic member/observer.

This was necessarily as it would take us quite a long time
to have a fully functional profiles. At that time we expect
to remove this field, together with the legacy tests.