WASM: Upgrade the WASM PVM when migrating to proto alpha
Prior to this patch, only newly originated smart rollups would benefit
from improvements brought by a new protocol. Now, when the WASM PVM
receives a `Protocol_migration` internal messages, it decodes its
payload in order to check if it needs to upgrade to a new revision.

This way, already existing smart rollups can benefit from new features
and improvements added in a more recent protocol.

This patch also introduces a test to check the consistency between the
naming scheme of the protocols, which have to be defined in two
places: `proto_*/lib_protocol/raw_context.ml` and
`lib_scoru_wasm/constants.ml`. Some instructions to update the test
when a new protocol is snapshot are also provided, to make the life of
developers doing the snapshot easier.