WASM: Parameterize `stack_size_limit` instead of hardcoding it to 300
The callstack in the lib_webassembly interpreter is limited to 300,
where 300 is hardcoded at the creation of the config parameter driving
the execution of a webassembly code.

With this patch, we turn this into a parameter that the caller can
modify.

The value chosen by the PVM remains 300 for now, for backward
compatibility reasons. Once the notion of WASM PVM revision is merged,
it will become possible to modify this parameter and makes it more
permissive.

Gitlab-Issue: https://gitlab.com/tezos/tezos/-/issues/4888