Proto/Michelson: Allow the elaborator to use the unparser
To be able to normalize the code embedded in each lambda value, the
elaborator needs to reuses the unparser.

The unparser was already indirectly dependent on the elaborator,
through a module functor.

Hence, this change introduces an (indirect) mutual dependency between
the elaborator and the unparser.

This is achieved thanks to a stratified definition of two functionals
finally applied together on the elaborator side.

Signed-off-by: Yann Regis-Gianas <yann@nomadic-labs.com>
Co-Authored-by: Raphaël Cauderlier <raphael.cauderlier@nomadic-labs.com>