WASM: Refactor WASM PVM versioning
Prior to this patch, the versioning of the WASM PVM was hard-coded in
the `initial_tree` version.

There is now several compliant use-case to introduce a newer version
of the WASM PVM, with refactoring of the existing host functions,
introduction of new host functions, and parameterization of the
maximum call depth of the WASM interpreter.

This patch is a first step towards introducing a new version to
welcome these breaking changes.