Proto: Inject an internal message when migrating to proto_alpha
This patch introduces a new kind of internal message, intended to be
injected by the economic protocol when a protocol migration happens.

The purpose is twofold: 1. allow kernels to know a protocol upgrade
happened, and 2. allow the PVM to upgrade itself to new
functionalities when it happens.

This patch only deals with purpose 1., that is the injection of the
message, already allowing kernels to interpret it as they see fit.

Co-authored-by: Valentin Chaboche <valentin.chaboche@nomadic-labs.com>