WASM: Introduce the host functions registry builder
The main motivation for this MR is to prepare the introduction of
per-revision host functions in the WASM PVM.

The newly introduced [builder] type allows to incrementally constructs
a [registry], and share a common base between two registries. This
should makes the use of a ad-hoc [write_debug] implementation less
demanding, performance-wise.