SCORU/Node: move the playing of refutation games to workers
The module [Refutation_player] defines a worker group that will be
handled by the refutation coordinator. A new refutation player will
be spawned for each new conflict that the node has to defend.

A [Refutation_player] can be initialized on a fresh conflict, and
therefore tasked with playing the opening move, or for an ongoing game.

This commit also adds a component to the rollup node for coordinating
refutation game players (as introduced in the previous commit).