Proto: Fix await_protocol_activation that does not wait anymore
The last release of resto has fixed a bug in which opening a stream
for reading was blocking as long as the first element was not received.
Unfortunately, the await_protocol_activation was built on top of this
side effect blocking semantic (considering the bug as a feature).
This patch fixes the problem in every concerned protocol and now waits
for the first received element explicitely.