DAL/crypto: fix initialisation_parameters_from_files
The size of the mapped file in bytes was incorrect.
It was equal to the number of elements in the file, but
not multiplied by the size in bytes of an element.

Here we simply remove the size of the mapped file
as Lwt_bytes.map_file will raise an error if the
supplied file size is greater than its actual file size.