WASM: [Host_funcs.lookup] per revision of the PVM
Before this patch, only the host funcs registry is specialized for
each WASM PVM revision. The [lookup] function is not.

This is actually bad, because during the linking phase, the [lookup]
function is used to “translate” the host function global name to their
implementation name.

Modifying this function to unconditionally add a new lookup would
therefore change the behavior of the WASM PVM: in a revision where the
new host function is missing, it will eventually fails, but not at the
same tick.

This patches fixes this issue, as a preliminary step towards adding a
new host function to the WASM PVM.