Proto/Michelson: always check annotations first in parse_ty
Michelson types can be annotated with at most one type annotation (and
no var nor field annotation except on subtypes of `pair`, `or` and
legacy `option`). Before this commit, this check was performed either
at the beginning of each case of parse_ty (for sapling types) or at
the end (for everything else). This commit consistently moves these
annotation checks at the beginning of each case of parse_ty.