Tezt/records: fail on HTTP errors
[dune exec tezt/records/update.exe -- -a from=last-merged-pipeline]
does not fail when the records of the last merged pipeline have been
removed by Gitlab. This is since, although GitLab returns a 404 error,
curl does not fail by default on HTTP errors. Here, we pass [--fail]
to curl, so that the script will fail when the artifacts are
removed.

Consequently, the job [tezt:fetch-records] will also fail, and the
fall back stored in the repo will be used instead.