P2p: Fix the maintainance triggering on new connections
This bug was triggering the maintainance when a new connection
is instantiated and the resulting number of connections equals
the maximum connections allowed by the node configuration.

This was not consistent with the checks done by the maintainance
worker that, when awaken, was ensuring that the number of active
connections was strictly superior than the maximum connections
allowed by the configuration before running the maintenance.

The trigger was used for nothing and it caused some side effects
like running a swap, etc...