Client/Michelson: fix numbering of literal scripts
This commit changes the way literal scripts are numbered in the `hash
script` and `typecheck script` commands. Instead of numbering all
scripts (literal or not), we now only increment the counter of literal
scripts once per literal script.

This only affects the output when both paths and literal scripts are
given. On `octez-client typecheck script some_path.tz "<first literal
script>" "<second literal script>"`, the literal scripts are now
numbered 1 and 2 instead of 2 and 3.