EVM/Execution: store accounts values as little endian
The reason for that is because cli likes eth-cli are decoding u256 as
little endian values, which means that an account that holds 9999 ETH,
will hold ~1.518E+72 ETH (when decoding as big endian) which is absurd.
As a side result, some of our E2E tests are failing as well.
Note that it doesn't change anything from evm-execution's perspective.