Proto/Michelson: add the UNIT instruction to the IR
The UNIT instruction is currently handled very specifically; it is a
kind of macro which is expanded to `PUSH unit Unit` at elaboration
time.

This commit promotes it to a real instruction by adding a case in the
GADT. The only observable effect is a slightly reduced footprint in
the contract cache for scripts using this instruction.