PlonK: commit RC_Z with wires
Note that this has a lot of side effects on aPlonK; in particular, Z_RC & Z_RC_perm are not batched together anymore.
Also, since Z_RC are now batched with wires, and since we need the next evaluation for Z_RC, this means if there is a range-check in a circuit, there will be a "next wire evaluation".