Shell: rename prechecked to validated
This commit is just a search/replace of all occurrences of
"prechecked" to "validated" in mempool files, except for
[Shell_metrics.Mempool.set_prechecked_collector] in
prevalidator_internal.ml because renaming metrics is more complex.

We use the term "validated" rather than "valid" to be able to search
for it in mempool-related files without getting overwhelmed with
instances of "prevalidator", "validation", "invalid", etc.