Tezt: Make the [Dac_node] module compatible with [Client.endpoint]
We follow the approach used by the Client module to not assume the
DAC node is communicating with an Octez node managed by Tezt. This
generalizes the Dac_node module, and makes it possible to [run]
a [Dac_node.t] process against a public endpoint like the one of
Mondaynet for instance.