Proto/AI: make the type of the launch_cycle context field optional
Until now we gave a special meaning to the absence of the
adaptive_inflation_launch_cycle context key. This sometimes lead to
several read attempts of the key during the application of a single
block, which the proxy server does not appreciate.

This commit changes the type of the key to `Cycle.t option` which
means that we are more explicit about the special `None` value.