CI/Tezt: store all logs
Since Tezt 3.1.0, logs are separated when executed with [-j X] where [X >
1]. With [--log-file tezt.log], report messages (success, failure),
are written to [tezt.log], where as log messages from the tests
running in process N are written to [tezt-N.log]

We use [-j 4] in the CI but we only store [tezt.log] as an
artifact. This MR changes to configuration to store all the tezt log
files, including those of the worker processes.