Proto/Michelson: simplify `DIP n`
Instead of using `PUSH` instructions to restore the stack prefix, we
rely on the `KUndip` continuation.

A side-effect of this commit is a simplification of the logging of
`DIP n` which is made more regular. Before this commit, consuming the
prefix was done in one logging step but restoring it was done in
several steps; now consuming and restoring the prefix are done in one
logging step each.