Proto/staking: refactor computation of stakes from snapshot
To set the distribution of delegates, we need to iterate over all
active delegates in the selected snapshot and compute for each of them
their stake.

In this commit, we extract the computation of the stake of each
delegate from the iteration because we want to expose this computation
in some RPCs.